refactor: rename sub-module azauth to authopts
