+++ LICENSE-THIRD-PARTY
@@ -3,11 +3,11 @@ Direct:
 * go-cmp (github.com/google/go-cmp) - BSD 3-Clause "New" or "Revised" License
 * sync (golang.org/x/sync) - BSD-3-Clause
 
-azauth:
+authopts:
 
 Direct:
 
-* azcore (github.com/Azure/azure-sdk-for-go/sdk/azcore) - MIT License (azauth)
+* azcore (github.com/Azure/azure-sdk-for-go/sdk/azcore) - MIT License
 * go-cmp (github.com/google/go-cmp) - BSD 3-Clause "New" or "Revised" License
 
 Indirect:
