+++ authopts/LICENSE-THIRD-PARTY
@@ -1,6 +1,6 @@
 Direct:
 
-* azcore (github.com/Azure/azure-sdk-for-go/sdk/azcore) - MIT License (azauth)
+* azcore (github.com/Azure/azure-sdk-for-go/sdk/azcore) - MIT License
 * go-cmp (github.com/google/go-cmp) - BSD 3-Clause "New" or "Revised" License
 
 Indirect:
