+++ authopts/authopts.go
@@ -1,4 +1,4 @@
-package azauth
+package authopts
 
 import (
 	"context"
@@ -25,7 +25,7 @@ type credential struct {
 // Token wraps around the TokenCredential method to satisfy the azcfg.Credential interface.
 func (c credential) Token(ctx context.Context) (auth.Token, error) {
 	token, err := c.TokenCredential.GetToken(ctx, policy.TokenRequestOptions{
-		Scopes: auth.Scopes,
+		Scopes: []string{"https://vault.azure.net/.default"},
 	})
 	if err != nil {
 		return auth.Token{}, err
